+++ src/cqterrain/__init__.py
@@ -3,6 +3,5 @@ from .Ladder import Ladder
 from .Board import Board
 
 # Methods
-from .stairs import stairs
 from .support import support
 from .obelisk import obelisk