+++ src/cqterrain/material/__init__.py
@@ -1,0 +1,1 @@
+from .stone import make_stones