+++ src/cqterrain/stairs/__init__.py
@@ -1,0 +1,1 @@
+from .stairs import stairs